Apply queueCapacity to pipelineDest as well as pipelineSource.
If the queue capacity is applied only to the pipelineSource queue but not the pipelineDest queue, we can run into problems when the Worker polls at a faster rate than the Indexer (e.g. if the pipeline is quick to execute but indexing is slow). Even if the pipelineSource queue reaches its capacity, the Worker will keep polling, creating enough space for one more document to be published. This can happen again and again, very rapidly, in a way where the pipelineSource queue is always at capacity (or capacity-1) but nevertheless the system is being flooded with new documents and the pipelineDest queue is growing without bound, which could use up all available memory. To prevent this situation, a simple fix is to apply a capacity to the pipelineDest queue as well as pipelineSource. A default capacity should always be applied if one isn't specified in the config, because we never want to risk an OME. If documents happen to be very large, the capacity may need to be reduced from the default of 100.