LC-135 - Initial pass at S3 Connector with basic functionality
The S3 connector allows you to connect to S3 and it publishes documents and Base64 encodes the content for passing through the system. The S3Connector should technically be named a VFSConnector because because it will also accept raw absolute and relative file paths as well as file:// schema. It needs tests and documentation still.