LC-135 - Modified test to ignore locally created .DS_Store files on OSX
When working locally in OSX, if you open a data directory with Finder, OSX can create .DS_Store files in the local directory that are found by connector. Normally include/exclude could be used to ensure are included / excluded, but the test it causes issues for is the test that makes sure all files are found in data subdirectory if there ARE no include/exclude filters. Normally there could be a file extension include filter and 0 or more exclude patterns to make the connector avoid unwanted directories / data