HybridKafka test had several places where we did a Thread.sleep() to wait for documents to be processed, but we had no way to guarantee that any particular sleep duration would be long enough, so the there was 1) possibly excessive/unnecessary time spent sleeping, 2) a possibility of transient failure.
Here we introduce an atomic counter that can be shared by all indexers running in a JVM. When an indexer sends an index event or callback, this counter is incremented. In test code, we can then wait for the counter to reach a desired value before shutting the indexers down.

An earlier approach was to add a config setting that would control the number of iterations that indexers and workers would go through before exiting their run loops. This doesn't work because when there are multiple indexers running, we don't know how much work each one will receive, so we don't know how many iterations to set as the cap. If the cap is too high and the work is exhausted, the indexer would begin blocking and the tests would hang. This could be addressed by introducing a configurable poll timeout, but it gets more complicated than the current solution.