rework test to account for problematic rebalance scenario
the scenario is where on worker commits an offset for a partition that's lower than the offset committed by another worker. This could happen in a rebalance scenario where
the first worker loses the partition to the second worker. If the first worker is delayed in committing its final offset for that partition (possibly because the corresponding batch took a long time to index), the first worker's offset could overwrite the second worker's offest. Even though the first worker's offset is lower, it gets committed later in time. This means that the sum of current committed offsets might be lower than the total number of documents processed.